requirements.txt
requests
python-dotenv
web3
eth-account